feat: global refactor
