+++ internal/metrics/endpoints.go
@@ -2,15 +2,19 @@ package metrics
 
 import (
 	"bytes"
-	"encoding/json"
 	"fmt"
-	"github.com/google/uuid"
-	myJson "github.com/https-whoyan/dwellingPayload/pkg/json"
-	"github.com/https-whoyan/dwellingPayload/pkg/repository/postgres"
 	"io"
+	"os"
+
+	"encoding/json"
 	"log/slog"
 	"net/http"
-	"os"
+
+	myJson "github.com/imperatorofdwelling/Website-backend/pkg/json"
+
+	"github.com/imperatorofdwelling/Website-backend/pkg/repository/postgres"
+
+	"github.com/google/uuid"
 )
 
 const (
