+++ internal/models/payment.go
@@ -2,8 +2,9 @@ package models
 
 import (
 	"fmt"
-	"github.com/google/uuid"
 	"time"
+
+	"github.com/google/uuid"
 )
 
 type TransferHistory struct {
