--- pkg/repository/postgres/auth_postgres.go
@@ -1,31 +1,0 @@
-package postgres
-
-// Below not need functionality
-/*
-type AuthPostgres struct {
-	db *sqlx.DB
-}
-
-func NewAuthPostgres(db *sqlx.DB) *AuthPostgres {
-	return &AuthPostgres{db}
-}
-
-func (r *AuthPostgres) CreateUser(user User) (int, error) {
-	var id int
-	query := fmt.Sprintf("INSERT INTO %s (name, username, password_hash) VALUES ($1, $2, $3) RETURNING id", usersTable)
-	row := r.db.QueryRow(query, user.Name, user.Username, user.Password)
-	if err := row.Scan(&id); err != nil {
-		return 0, err
-	}
-
-	return id, nil
-}
-
-func (r *AuthPostgres) GetUser(username, password string) (User, error) {
-	var user User
-	query := fmt.Sprintf("SELECT id FROM %s WHERE username=$1 AND password_hash=$2", usersTable)
-	err := r.db.Get(&user, query, username, password)
-
-	return user, err
-}
-*/
