--- pkg/repository/postgres/db.go
@@ -1,20 +1,0 @@
-package postgres
-
-// Below not needed code
-/*
-type Authorization interface {
-	//CreateUser(user User) (int, error)
-	//GetUser(username, password string) (todo.User, error)
-}
-
-type Repository struct {
-	Authorization
-	// TODO: transaction ...
-}
-
-func NewRepository(db *sqlx.DB) *Repository {
-	return &Repository{
-		Authorization: NewAuthPostgres(db),
-	}
-}
-*/
